Check user/default config locations This code has NOT been tested. Re-did the configObj parsing to use validate(). This provides additional automated checking on user inputs, the ability to roll back to a default value, and check if a default value was used. This gives us extra flexibility down the road.
Modified the default settings file to a format compliant with validate()

For the settings themselves, we first check if the value being used is default or user supplied. If user supplied, we check that value to see if it is
a valid path. If it is not valid, or the default value was used, we then check to see if the default value is valid. If it is, we use it, if it is not,
we log an error and quit.